Require an Instructor role to do export
Instead of fobidding export to users who have a TA role, we now only
allow it to users who have the Instructor role, which is given to
examiners, course responsibles and teachers, but not to TA:s.

Note that we use `roles`, and not `ext_roles` here.

Checking ext_roles for `urn:lti:role:ims/lis/Instructor` might also be
ok (but not urn:lti:instrole:ims/lis/Instructor, as that is given to
TA:s also).

Also note that we may want to allow
`urn:lti:instrole:ims/lis/Administrator` here, but 1. I don't know
that we want that, and 2. that role is required to even log in to the
test canvas.